+++ vpr/test/test_vpr.cpp
@@ -113,6 +113,7 @@ TEST_CASE("read_arch_metadata", "[vpr]") {
 }
 
 TEST_CASE("read_rr_graph_metadata", "[vpr]") {
+    /* TODO: All the inode should use RRNodeId */
     int src_inode = -1;
     int sink_inode = -1;
     short switch_id = -1;
@@ -137,7 +138,7 @@ TEST_CASE("read_rr_graph_metadata", "[vpr]") {
 
         for (const RRNodeId& inode : device_ctx.rr_graph.nodes()) {
             if ((rr_graph.node_type(inode) == CHANX || rr_graph.node_type(inode) == CHANY) && rr_graph.num_edges(inode) > 0) {
-                src_inode = inode;
+                src_inode = size_t(inode);
                 break;
             }
         }
